Restructure tests [#164069470]
