+++ src/controllers/party.js
@@ -1,5 +1,5 @@
 import joi from 'joi';
-import { parties, offices } from '../database';
+import { parties } from '../database';
 
 const createParty = (req, res) => {
   const schema = {
@@ -30,39 +30,11 @@ const createParty = (req, res) => {
   });
 };
 
-
-const createOffice = (req, res) => {
-
-  const schema = {
-    type: joi.string()
-      .min(1)
-      .required(),
-    name: joi.string()
-      .min(1)
-      .required(),
-  };
-  const result1 = joi.validate(req.body, schema);
-
-  if (result1.error) {
-    return res.status(400).json({
-      status: 400,
-      message: result1.error.details[0].message,
-    });
-  }
-  const office = { id: offices.length + 1, ...req.body, createdOn: new Date() };
-
-  offices.push(office);
-  return res.status(201).json({
-    status: 201,
-    data: [office],
-  });
-};
-
-
 const getAllParties = (req, res) => res.status(200).json({
   status: 200,
   data: parties,
 });
+
 const getParty = (req, res) => {
   const { id } = req.params;
   const party = parties.find(p => p.id === parseInt(id, 10));
@@ -77,6 +49,7 @@ const getParty = (req, res) => {
     data: [party],
   });
 };
+
 const updateParty = (req, res) => {
   const { id } = req.params;
   const party = parties.find(p => p.id === parseInt(id, 10));
@@ -106,24 +79,6 @@ const deleteParty = (req, res) => {
     data: [party],
   });
 };
-const getAllOffices = (req, res) => res.status(200).json({
-  status: 200,
-  data: offices,
-});
-const getOffice = (req, res) => {
-  const { id } = req.params;
-  const office = offices.find(f => f.id === parseInt(id, 10));
-  if (!office) {
-    res.status(404).json({
-      status: 404,
-      message: 'the office with a given id does not exist',
-    });
-  }
-  return res.status(200).json({
-    status: 200,
-    data: [office],
-  });
-};
 
 export {
   createParty,
@@ -131,8 +86,4 @@ export {
   getParty,
   deleteParty,
   updateParty,
-  createOffice,
-  getAllOffices,
-  getOffice,
-
 };
