+++ src/tests/index.js
@@ -1,1 +1,2 @@
-import './parties.test'
+import './parties.test';
+import './offices.test';
