+++ src/tests/parties.test.js
@@ -1,6 +1,5 @@
 import chai from 'chai';
 import chaiHttp from 'chai-http';
-import db from '../database';
 import app from '../app';
 
 const {
@@ -51,47 +50,27 @@ describe('Parties', () => {
         });
     });
   });
-});
-/* Offices */
-describe('Offices', () => {
-  describe('POST /api/v1/offices', () => {
-    // test 2
-    it('should return a created office', (done) => {
+  describe('PATCH /api/v1/parties/:id/:name', ()=>{
+    it('should update the name of a party', (done)=>{
       chai.request(app)
-        .post('/api/v1/offices')
-        .send({
-          name: 'President',
-          type: 'Presidential',
-        })
-        .end((err, res) => {
-          expect(res.status).to.equal(201);
-          expect(res.body.data.length).to.be.above(0);
-          done();
-        });
+      .patch('api/v1/parties/1/PL')
+      .end(err, res)=>{
+        expect(res.status).to.equal(200);
+        expect(res.body.data.length).to.be.above(0);
+        done();
+      });
     });
   });
-  describe('GET /api/v1/offices', () => {
-    // test 2
-    it('should return a list of offices', (done) => {
+  describe('DELETE /api/v1/parties/:id', () => {
+    // test 1
+    it('should return the deleted party', (done) => {
       chai.request(app)
-        .get('/api/v1/offices')
+        .delete('/api/v1/parties/1')
         .end((err, res) => {
           expect(res.status).to.equal(200);
           expect(res.body.data.length).to.be.above(0);
           done();
         });
     });
-  });
-});
-describe('GET /api/v1/offices/:id', () => {
-  // test 3
-  it('should return an office', (done) => {
-    chai.request(app)
-      .get('/api/v1/offices/1')
-      .end((err, res) => {
-        expect(res.status).to.equal(200);
-        expect(res.body.data.length).to.be.above(0);
-        done();
-      });
   });
 });
