--- tests/testData.js
@@ -1,41 +1,0 @@
-const dummyParties1 = [
-  {
-        "name":"RPF",
-        "hqAddress":"Kigali",
-        "logoUrl":"weruuu"
-    }
-];
-  const dummyParties2 =[
-    {
-    id: 2,
-    name: 'RPF',
-    hqAddress: 'Kigali',
-    logoUrl: 'adnckadcnjad.com'
-  }
-];
-
-const dummyOffice = [
-  {
-    type: 'federal',
-    name: 'federal bureau',
-  }
-];
-const dummyOffices = [
-  {
-    id: 1,
-    type: 'federal',
-    name: 'federal bureau',
-  },
-
-  {
-    id: 2,
-    type: 'federal',
-    name: 'federal bureau',
-  },
-  {
-    id: 3,
-    type: 'federal',
-    name: 'federal bureau',
-  },
-];
-export { dummyParties1,dummyParties2, dummyOffices, dummyOffice };
